Move view up/down when keyboard is displayed
Implemented ability to move the UIView up or down depending on when the
keyboard is displayed or hidden.
INCOMPLETE.

This needs lot of work. As the size of the keyboard is required to
exactly determine where the keyboard is displayed.

Also currently it only works for the top two textbooks. The 3rd textbox
is still hidden under the keyboard.